A primitive deferred view maintenance
This change is creating a primitive deferred view maintenance mode.
Register a query and updates the cache on that. This implementation
relies on H2 as an in-memory SQL database, so we can compute the changes
for each event and update the cache accordingly.

This probably has limitations in the different SQL dialects, data types,
and definitely a bunch of unknown unknowns. We use MySQL and our simple
customer and car model example to make things work.